Use post read control on add operations
By adding the post read control, we gain two benefits. First,
the immediate read that is performed after the add is no longer
necessary; saving a network roundtrip. Second, if an entry's DN
changes as part of the add procedure (ex. via the IPA UUID plugin)
this change is picked up automatically.